+++ NEWS.md
@@ -3,6 +3,8 @@ Red Panda C++ Version 1.3
   - enhancement: don't parse all openned files when start up
   - enhancement: don't parse files when close all and exit
   - change: reduce time intervals for selection by mouse
+  - enhancement: change orders of the problems in the problem set panel by drag&drop
+  - enhancement: change orders of the problem cases in the problem panel by drag&drop
 
 Red Panda C++ Version 1.2